Add package layer upper commands to be consistent with web.xml CommandAccessFilter
