+++ src/main/java/com/a_smart_cookie/controller/command/CommandContext.java
@@ -1,5 +1,15 @@
 package com.a_smart_cookie.controller.command;
 
+import com.a_smart_cookie.controller.command.common.LogoutCommand;
+import com.a_smart_cookie.controller.command.guest.LoginCommand;
+import com.a_smart_cookie.controller.command.guest.RegistrationCommand;
+import com.a_smart_cookie.controller.command.guest.SignInPageCommand;
+import com.a_smart_cookie.controller.command.guest.SignUpPageCommand;
+import com.a_smart_cookie.controller.command.shared.CatalogCommand;
+import com.a_smart_cookie.controller.command.subscriber.PaymentCommand;
+import com.a_smart_cookie.controller.command.subscriber.SubscribeCommand;
+import com.a_smart_cookie.controller.command.subscriber.SubscriptionsCommand;
+import com.a_smart_cookie.controller.command.subscriber.UnsubscribeCommand;
 import org.apache.log4j.Logger;
 
 import java.util.HashMap;
