+++ src/main/java/com/a_smart_cookie/controller/command/guest/SignUpPageCommand.java
@@ -1,5 +1,6 @@
-package com.a_smart_cookie.controller.command;
+package com.a_smart_cookie.controller.command.guest;
 
+import com.a_smart_cookie.controller.command.Command;
 import com.a_smart_cookie.controller.route.HttpHandlerType;
 import com.a_smart_cookie.controller.route.HttpPath;
 import com.a_smart_cookie.controller.route.WebPath;
@@ -8,17 +9,19 @@ import org.apache.log4j.Logger;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class SignInPageCommand extends Command {
 
-	private static final long serialVersionUID = -7839366578058993131L;
+public class SignUpPageCommand extends Command {
 
-	private static final Logger LOG = Logger.getLogger(SignInPageCommand.class);
+	private static final long serialVersionUID = -6062947839872352840L;
+
+	private static final Logger LOG = Logger.getLogger(SignUpPageCommand.class);
 
 	@Override
 	public HttpPath execute(HttpServletRequest request, HttpServletResponse response) {
 		LOG.debug("Command starts");
 
 		LOG.debug("Command finished");
-		return new HttpPath(WebPath.Page.SIGN_IN, HttpHandlerType.FORWARD);
+		return new HttpPath(WebPath.Page.SIGN_UP, HttpHandlerType.FORWARD);
 	}
+
 }
