+++ src/main/java/com/a_smart_cookie/controller/command/shared/CatalogCommand.java
@@ -1,5 +1,6 @@
-package com.a_smart_cookie.controller.command;
+package com.a_smart_cookie.controller.command.shared;
 
+import com.a_smart_cookie.controller.command.Command;
 import com.a_smart_cookie.controller.route.HttpHandlerType;
 import com.a_smart_cookie.controller.route.HttpPath;
 import com.a_smart_cookie.controller.route.WebPath;
